Make wheel slimmer

+++ sematic/tests/integration/test_pip_install.sh
@@ -16,7 +16,31 @@ pwd
 
 source ./$VENV_NAME/bin/activate
 
+WHEEL_PATH=$(ls bazel-bin/sematic/sematic-*.whl)
+
+if test -f "$WHEEL_PATH"; then
+    echo "Error: Wheel not found"
+    exit 1
+fi
+
+N_MB_SIZE_LIMIT=10
+WHEEL_SIZE_MB=$(python3 -c "import os; print(int(os.path.getsize('$WHEEL_PATH') / 2**20))")
+
+if (( WHEEL_SIZE_BYTES > N_BYTE_SIZE_LIMIT )); then
+    echo "Error: Wheel bigger than $N_MB_SIZE_LIMIT Mb. Size: $WHEEL_SIZE_MB Mb"
+    exit 1
+else
+    echo "Wheel is $WHEEL_SIZE_MB Mb"
+fi
+
 pip install bazel-bin/sematic/sematic-*.whl
+
+if $? ; then
+    echo "Pip install succeeded"
+else
+    echo "Pip install failed"
+    exit 1
+fi
 
 deactivate
